test: change the location of dummy samples
